ai/requirements.txt
transformers
torch
numpy==1.21.6
pandas
scikit-learn
textstat
python-dotenv
sagemaker>=2.48.0
nltk
spacy==3.5.0